dev: Refactoring & Improvements
- Refactor data schema: changed all primary (PK) and foreign (FK) key types from long to int across all entities, navigation properties, and DTOs.
- Add CompanyData entity, DTOs, service, and controller for public/admin company info management.
- Move/refactor Vendor and WarehouseDelivery controllers to Admin namespace with full CRUD operations and XML documentation.
- Implement AuthController for registration and login using JWT.
- Update all service interfaces and implementations to use int IDs; add XML documentation to all service methods.
- Add paginated public item catalog and detail endpoints.
- Add comprehensive XML documentation to all DTOs and entities.
- Reset database migrations: remove legacy migrations and add initial migration including the CompanyData table.
- Configure ICompanyDataService in DI; update connection string settings and Docker Compose for SQL Server.
- General cleanup: improve code comments, summaries, and remove obsolete files.

+++ Backend/Controllers/OrderController.cs
@@ -4,17 +4,29 @@ using Microsoft.AspNetCore.Mvc;
 
 namespace Backend.Controllers;
 
+/// <summary>
+/// Manages customer orders and order operations.
+/// Provides CRUD endpoints for managing orders including creation, retrieval, updating, and deletion.
+/// </summary>
 [Route("api/[controller]")]
 [ApiController]
 public class OrderController : ControllerBase
 {
     private readonly IOrderService _service;
 
+    /// <summary>
+    /// Initializes a new instance of the <see cref="OrderController"/> class.
+    /// </summary>
+    /// <param name="service">The order service for handling order operations.</param>
     public OrderController(IOrderService service)
     {
         _service = service;
     }
 
+    /// <summary>
+    /// Retrieves all orders.
+    /// </summary>
+    /// <returns>A list of all orders in the system.</returns>
     [HttpGet]
     public async Task<IActionResult> GetAll()
     {
@@ -22,13 +34,23 @@ public class OrderController : ControllerBase
     }
 
 
+    /// <summary>
+    /// Retrieves a specific order by its ID.
+    /// </summary>
+    /// <param name="id">The ID of the order to retrieve.</param>
+    /// <returns>The order if found; otherwise, a 404 Not Found response.</returns>
     [HttpGet("{id}")]
-    public async Task<IActionResult> GetById(long id)
+    public async Task<IActionResult> GetById(int id)
     {
         var vendor = await _service.GetOrderByIdAsync(id);
         return vendor is null ? NotFound() : Ok(vendor);
     }
 
+    /// <summary>
+    /// Creates a new order.
+    /// </summary>
+    /// <param name="order">The order to create.</param>
+    /// <returns>The created order with an assigned ID.</returns>
     [HttpPost]
     public async Task<IActionResult> Create(Order order)
     {
@@ -36,6 +58,11 @@ public class OrderController : ControllerBase
     }
 
 
+    /// <summary>
+    /// Updates an existing order.
+    /// </summary>
+    /// <param name="order">The order with updated information.</param>
+    /// <returns>The updated order.</returns>
     [HttpPut]
     public async Task<IActionResult> Update(Order order)
     { 
@@ -43,8 +70,13 @@ public class OrderController : ControllerBase
     }
 
 
+    /// <summary>
+    /// Deletes an order by its ID.
+    /// </summary>
+    /// <param name="id">The ID of the order to delete.</param>
+    /// <returns>A boolean indicating whether the deletion was successful.</returns>
     [HttpDelete("{id}")]
-    public async Task<IActionResult> Delete(long id)
+    public async Task<IActionResult> Delete(int id)
     { 
         return Ok(await _service.DeleteOrderAsync(id)); 
     }
